requirements.txt
pyserial==3.5
pygame==2.5.2
requests==2.31.0
